requirements.txt
APScheduler==3.6.3
astroid==2.3.3
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.1
Django==2.2.7
django-chartjs==1.5.0
django-filter==2.2.0
djangorestframework==3.11.0
factory-boy==2.12.0
Faker==2.0.4
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
nltk==3.4.5
oauthlib==3.1.0
psycopg2==2.8.4
psycopg2-binary==2.8.4
pylint==2.4.4
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
sqlparse==0.3.0
text-unidecode==1.3
textblob==0.15.3
tweepy==3.8.0
typed-ast==1.4.0
tzlocal==2.0.0
urllib3==1.25.7
virtualenv==16.7.5
wrapt==1.11.2
